Removing grid.py from Parcels v4
