+++ parcels/__init__.py
@@ -5,7 +5,6 @@ __version__ = version
 from parcels.application_kernels import *
 from parcels.field import *
 from parcels.fieldset import *
-from parcels.grid import *
 from parcels.interaction import *
 from parcels.kernel import *
 from parcels.particle import *
