+++ tests/v4/test_xgrid.py
@@ -7,7 +7,6 @@ from numpy.testing import assert_allclose
 
 from parcels import xgcm
 from parcels._datasets.structured.generic import T, X, Y, Z, datasets
-from parcels.grid import Grid as OldGrid
 from parcels.xgrid import XGrid, _search_1d_array
 
 GridTestCase = namedtuple("GridTestCase", ["Grid", "attr", "expected"])
@@ -38,36 +37,6 @@ def assert_equal(actual, expected):
 def test_xgrid_properties_ground_truth(ds, attr, expected):
     grid = XGrid(xgcm.Grid(ds, periodic=False))
     actual = getattr(grid, attr)
-    assert_equal(actual, expected)
-
-
-@pytest.mark.parametrize(
-    "attr",
-    [
-        "lon",
-        "lat",
-        "depth",
-        "time",
-        "xdim",
-        "ydim",
-        "zdim",
-        "tdim",
-        "_gtype",
-    ],
-)
-@pytest.mark.parametrize("ds", [pytest.param(ds, id=key) for key, ds in datasets.items()])
-def test_xgrid_against_old(ds, attr):
-    grid = XGrid(xgcm.Grid(ds, periodic=False))
-
-    old_grid = OldGrid.create_grid(
-        lon=ds.lon.values,
-        lat=ds.lat.values,
-        depth=ds.depth.values,
-        time=ds.time.values.astype("float64") / 1e9,
-        mesh="spherical",
-    )
-    actual = getattr(grid, attr)
-    expected = getattr(old_grid, attr)
     assert_equal(actual, expected)
